Cleanup other MDM commands tests
